Reduce build times by refactoring JSDOMGlobalObject.h https://bugs.webkit.org/show_bug.cgi?id=251570 rdar://104949069
Reviewed by Ryosuke Niwa.

Move inline methods into JSDOMGlobalObjectInlines.h, and add that include to all source files
where those inlines are used.

Allow JSBuiltinInternalFunctions to be forward declared by putting the m_builtinInternalFunctions
in a UniqueRef rather than a bare object.

Fix some drive-by build errors revealed by removing inline headers from JSDOMGlobalObject.h.

Canonical link: https://commits.webkit.org/259903@main
