+++ Source/WebCore/bindings/js/JSNavigatorCustom.cpp
@@ -26,6 +26,7 @@
 #include "config.h"
 #include "JSNavigator.h"
 
+#include "WebCoreJSBuiltinInternals.h"
 #include "WebCoreJSClientData.h"
 #include "WebCoreOpaqueRoot.h"
 #include <JavaScriptCore/CatchScope.h>
