+++ Source/WebCore/inspector/CommandLineAPIModule.cpp
@@ -27,6 +27,7 @@
 #include "CommandLineAPIModule.h"
 
 #include "JSDOMGlobalObject.h"
+#include "WebCoreJSBuiltinInternals.h"
 #include "WebInjectedScriptManager.h"
 #include <JavaScriptCore/HeapInlines.h>
 #include <JavaScriptCore/InjectedScript.h>
